+ Motion and a few Examples
(The idea of motion is that it's a container for relative vector positions. After initialization, you can call Step on the motion to get the next relative position and apply it to your entity. I intend to implement this not just for buildings, but for humanoids as well as UI buttons as well and maybe fun little easter egg animations like clicking on bushes and such)